renamed folders and added initial intruder states

+++ archive/main.py
@@ -4,7 +4,7 @@ from visualize3d import visualization
 def main():
     # visualization("short" or "long", number of intruders,
     #                   "body" or "static" inertial frame)
-    graph = visualization("long",20,"body")
+    graph = visualization("short",20,"body")
 
     while(True):
         graph.update_graph()
